chore: Refactored code related to variant images into a component
