+++ web-src/src/components/FavoriteVariantCard.js
@@ -10,8 +10,7 @@
  * governing permissions and limitations under the License.
  */
 import {
-  Button, ActionButton, Flex, Tooltip, TooltipTrigger, View, ContextualHelp, Heading, Content, Image, ProgressCircle,
-  MenuTrigger, Menu, Item, Text, DialogContainer, AlertDialog, Divider,
+  Button, ActionButton, Flex, Tooltip, TooltipTrigger, View, ContextualHelp, Heading, Content, ProgressCircle, Divider,
 } from '@adobe/react-spectrum';
 import React, {
   useCallback, useState,
@@ -23,21 +22,14 @@ import { useToggleFavorite } from '../state/ToggleFavoriteHook.js';
 import { useVariantImages } from '../state/VariantImagesHook.js';
 import { useApplicationContext } from './ApplicationProvider.js';
 import { useShellContext } from './ShellProvider.js';
-import {
-  toHTML, toText,
-} from '../helpers/PromptExporter.js';
-import {
-  generateImagePrompt, copyImageToClipboard, copyImageToClipboardLegacy, downloadImage as handleDownloadImage,
-} from '../helpers/ImageHelper.js';
+import { toHTML, toText } from '../helpers/PromptExporter.js';
+import { generateImagePrompt } from '../helpers/ImageHelper.js';
 import { sampleRUM } from '../rum.js';
-import { ImageViewer } from './ImageViewer.js';
+import { VariantImagesView } from './VariantImagesView.js';
 
 import CopyOutlineIcon from '../icons/CopyOutlineIcon.js';
-import EditIcon from '../icons/EditIcon.js';
 import DeleteOutlineIcon from '../icons/DeleteOutlineIcon.js';
 import GenAIIcon from '../icons/GenAIIcon.js';
-import DownloadIcon from '../icons/DownloadIcon.js';
-import MoreIcon from '../icons/MoreIcon.js';
 
 const styles = {
   card: css`
@@ -49,45 +41,15 @@ const styles = {
     padding: 10px;
     min-height: 100px;
   `,
-  variantThumbImages: css`
-    display: flex;
-    flex-direction: row;
-    gap: 10px;
-    justify-content: left;
-    width: 100%;
-    flex-wrap: wrap;
-    padding: 10px;
-  `,
-  variantThumbImage: css`
-    width: 144px;
-    height: 144px;
-    border-radius: 8px;
-  `,
 };
 
 export function FavoriteVariantCard({ variant, ...props }) {
   const { firefallService, expressSDKService } = useApplicationContext();
   const { user, isExpressAuthorized } = useShellContext();
   const toggleFavorite = useToggleFavorite();
-  const {
-    variantImages, addImageToVariant, replaceImageFromVariant, deleteImageFromVariant,
-  } = useVariantImages();
+  const { addImageToVariant } = useVariantImages();
 
   const [imagePromptProgress, setImagePromptProgress] = useState(false);
-  const [isImageViewerOpen, setIsImageViewerOpen] = useState(false);
-  const [imageViewerIndex, setImageViewerIndex] = useState(0);
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
-  const [selectedImageIndex, setSelectedImageIndex] = useState(null);
-  const [isMoreMenuOpen, setIsMoreMenuOpen] = useState(false);
-
-  const handleCopyImage = useCallback((base64Image) => {
-    if (navigator.userAgent.toLowerCase().indexOf('firefox') > -1) {
-      copyImageToClipboardLegacy(base64Image);
-    } else {
-      copyImageToClipboard(base64Image, 'image/png');
-    }
-    ToastQueue.positive('Copied image to clipboard', { timeout: 1000 });
-  }, []);
 
   const handleGenerateImage = useCallback(async (imagePrompt) => {
     const onPublish = (publishParams) => {
@@ -110,37 +72,6 @@ export function FavoriteVariantCard({ variant, ...props }) {
     );
   }, [expressSDKService, user, variant]);
 
-  const handleEditGenerateImage = useCallback(async (index) => {
-    const onPublish = (publishParams) => {
-      replaceImageFromVariant(variant.id, index, publishParams.asset[0].data);
-    };
-    const assetData = variantImages[variant.id][index];
-
-    await expressSDKService.handleImageOperation(
-      'editImage',
-      {
-        outputParams: {
-          outputType: 'base64',
-        },
-        inputParams: {
-          asset: {
-            data: assetData,
-            type: 'image',
-            dataType: 'base64',
-          },
-        },
-        callbacks: {
-          onPublish,
-        },
-      },
-    );
-  }, [expressSDKService, user, variant, variantImages]);
-
-  const handleImageViewerOpen = useCallback((index) => {
-    setImageViewerIndex(index);
-    setIsImageViewerOpen(true);
-  }, []);
-
   const handleGenerateImagePrompt = useCallback(() => {
     setImagePromptProgress(true);
     generateImagePrompt(firefallService, variant)
@@ -160,112 +91,9 @@ export function FavoriteVariantCard({ variant, ...props }) {
       initial={{ opacity: 0, scale: 0.9 }}
       animate={{ opacity: 1, scale: 1 }}
       transition={{ ease: 'easeIn', duration: 0.3 }}>
-      <div {...props}
-        className={styles.card}
-        onClick={() => setSelectedImageIndex(null)}>
-        <ImageViewer
-          images={variantImages[variant.id]}
-          index={imageViewerIndex}
-          onIndexChange={setImageViewerIndex}
-          open={isImageViewerOpen}
-          onClose={() => setIsImageViewerOpen(false)}
-          onEdit={(index) => {
-            setIsImageViewerOpen(false);
-            handleEditGenerateImage(index);
-          }}
-          onCopy={(index) => {
-            handleCopyImage(variantImages[variant.id][index]);
-          }}
-          onDownload={(index) => handleDownloadImage(variantImages[variant.id][index])}
-        />
-        <DialogContainer onDismiss={() => setIsDeleteDialogOpen(false)}>
-          {isDeleteDialogOpen && (
-            <AlertDialog
-              variant="destructive"
-              title="Delete Image"
-              primaryActionLabel="Delete"
-              cancelLabel="Cancel"
-              onPrimaryAction={() => {
-                deleteImageFromVariant(variant.id, selectedImageIndex);
-                setIsDeleteDialogOpen(false);
-              }}
-              onCancelAction={() => {
-                console.log("You've canceled the delete action");
-                setIsDeleteDialogOpen(false);
-              }}>
-              This will permanently delete the image. Continue?
-            </AlertDialog>
-          )}
-        </DialogContainer>
+      <div {...props} className={styles.card}>
         <div className={styles.variant} dangerouslySetInnerHTML={{ __html: toHTML(variant.content) }} />
-        {variantImages[variant.id]
-          && <div className={styles.variantThumbImages}>
-            {variantImages[variant.id].map((base64Image, index) => {
-              return (
-                <div key={index} className={'variant-image-wrapper'}>
-                  <div onClick={() => handleImageViewerOpen(index)}>
-                    <Image src={base64Image}
-                      objectFit={'cover'}
-                      UNSAFE_className={`${styles.variantThumbImage} ${(!isMoreMenuOpen || selectedImageIndex !== index) ? 'variant-thumb-image' : 'variant-thumb-image-selected'} hover-cursor-pointer`}
-                    />
-                  </div>
-                  <Flex direction={'row'} width={'100%'} gap={'size-100'} position={'absolute'} bottom={'size-100'} justifyContent={'center'}>
-                    <TooltipTrigger delay={0}>
-                      <Button
-                        variant='secondary'
-                        style='fill'
-                        UNSAFE_className={`${(!isMoreMenuOpen || selectedImageIndex !== index) && 'variant-image-button'} hover-cursor-pointer`}
-                        onPress={() => handleCopyImage(base64Image)}>
-                        <CopyOutlineIcon />
-                      </Button>
-                      <Tooltip>Copy Image</Tooltip>
-                    </TooltipTrigger>
-                    <TooltipTrigger delay={0}>
-                      <Button
-                        variant='secondary'
-                        style='fill'
-                        UNSAFE_className={`${(!isMoreMenuOpen || selectedImageIndex !== index) && 'variant-image-button'} hover-cursor-pointer`}
-                        onPress={() => handleEditGenerateImage(index)}>
-                        <EditIcon />
-                      </Button>
-                      <Tooltip>Edit</Tooltip>
-                    </TooltipTrigger>
-                    <TooltipTrigger delay={0}>
-                      <MenuTrigger onOpenChange={(isOpen) => {
-                        setIsMoreMenuOpen(isOpen);
-                        setSelectedImageIndex(index);
-                      }}>
-                        <Button
-                          variant='secondary'
-                          style='fill'
-                          UNSAFE_className={`${(!isMoreMenuOpen || selectedImageIndex !== index) && 'variant-image-button'} hover-cursor-pointer`}>
-                          <MoreIcon />
-                        </Button>
-                        <Menu width="size-1700" onAction={(key) => {
-                          if (key === 'download') {
-                            handleDownloadImage(base64Image);
-                          } else if (key === 'delete') {
-                            setIsDeleteDialogOpen(true);
-                          }
-                        }}>
-                          <Item key="download">
-                            <DownloadIcon UNSAFE_style={{ boxSizing: 'content-box' }} />
-                            <Text>Download</Text>
-                          </Item>
-                          <Item key="delete">
-                            <DeleteOutlineIcon UNSAFE_style={{ boxSizing: 'content-box' }} />
-                            <Text>Delete</Text>
-                          </Item>
-                        </Menu>
-                      </MenuTrigger>
-                      <Tooltip>More</Tooltip>
-                    </TooltipTrigger>
-                  </Flex>
-                </div>
-              );
-            })}
-          </div>
-        }
+        <VariantImagesView variant={variant} isFavorite={true} />
         <View marginTop={'10px'}>
           <Flex direction="row" justifyContent="left">
             <TooltipTrigger delay={0}>
